refactor: separate utility functions into lib modules
