+++ scripts/collect-translations.js
@@ -9,8 +9,10 @@
 
 const fs = require("fs");
 const path = require("path");
+const { resolvePackagePath, scanPackage } = require("./lib/translationUtils");
 
-const OUTPUT_DIR = process.env.I18N_OUTPUT_DIR || path.join("src", "translations");
+const OUTPUT_DIR =
+  process.env.I18N_OUTPUT_DIR || path.join("src", "translations");
 
 /**
  * MAIN ALGORITHM: Translation Collection
@@ -137,214 +139,5 @@ async function main() {
 main().catch(console.error);
 
 // =============================================================================
-// IMPLEMENTATION FUNCTIONS
+// Implementation functions in ./lib/translationUtils.js
 // =============================================================================
-
-/**
- * Find package in multiple fallback locations.
- */
-function resolvePackagePath(packageName) {
-  const locations = [
-    process.env.INVENIO_PACKAGES_DIR &&
-      path.join(process.env.INVENIO_PACKAGES_DIR, packageName),
-    path.resolve(__dirname, "../..", packageName),
-    path.resolve(__dirname, "..", packageName),
-  ].filter(Boolean);
-
-  return locations.find(fs.existsSync) || null;
-}
-
-/**
- * Parses a .po file using the i18next-conv library and converts it to translation format.
- */
-async function parsePoFile(poFilePath, packageName) {
-  try {
-    const { gettextToI18next } = await import("i18next-conv");
-    const poContent = fs.readFileSync(poFilePath, "utf8");
-
-    const result = await gettextToI18next("", poContent, {
-      skipUntranslated: true,
-      compatibilityJSON: "v4",
-    });
-
-    const translations = JSON.parse(result);
-    const output = {};
-
-    for (const [key, value] of Object.entries(translations)) {
-      if (key && value && typeof value === "string") {
-        const translationValue = value.trim() || key;
-        output[key] = translationValue;
-        output[`${packageName}:${key}`] = translationValue;
-      }
-    }
-
-    return output;
-  } catch (error) {
-    console.warn(`Failed to parse ${poFilePath}:`, error.message);
-    return {};
-  }
-}
-
-/**
- * Analyzes PO file for untranslated strings and quality issues.
- */
-
-/**
- * Creates a validation report for PO file analyzing untranslated strings and quality issues.
- *
- * @param {string} poFilePath - Path to the .po file
- * @param {string} packageName - Package name for context
- * @param {string} locale - Locale being analyzed
- * @returns {Object} Validation report
- */
-function createValidationReportOfPoFile(poFilePath, packageName, locale) {
-  try {
-    const poContent = fs.readFileSync(poFilePath, "utf8");
-    const lines = poContent.split("\n");
-
-    const issues = {
-      untranslated: [],
-      fuzzyTranslations: [],
-      obsoleteTranslations: [],
-    };
-
-    let currentMsgid = "";
-    let currentMsgstr = "";
-    let isFuzzy = false;
-    let isObsolete = false;
-    let inMsgid = false;
-    let inMsgstr = false;
-
-    for (const line of lines) {
-      const trimmed = line.trim();
-
-      // Check for flags
-      if (trimmed.startsWith("#,") && trimmed.includes("fuzzy")) {
-        isFuzzy = true;
-      }
-      if (trimmed.startsWith("#~")) {
-        isObsolete = true;
-      }
-
-      // Parse msgid and msgstr
-      if (trimmed.startsWith("msgid ")) {
-        currentMsgid = trimmed.substring(6).replace(/^"(.*)"$/, "$1");
-        inMsgid = true;
-        inMsgstr = false;
-      } else if (inMsgid && trimmed.startsWith('"') && trimmed.endsWith('"')) {
-        currentMsgid += trimmed.replace(/^"(.*)"$/, "$1");
-      } else if (trimmed.startsWith("msgstr ")) {
-        currentMsgstr = trimmed.substring(7).replace(/^"(.*)"$/, "$1");
-        inMsgid = false;
-        inMsgstr = true;
-      } else if (inMsgstr && trimmed.startsWith('"') && trimmed.endsWith('"')) {
-        currentMsgstr += trimmed.replace(/^"(.*)"$/, "$1");
-      } else if (trimmed === "" || trimmed.startsWith("msgid ")) {
-        // End of entry - process it
-        if (currentMsgid && currentMsgid !== "") {
-          if (isObsolete) {
-            issues.obsoleteTranslations.push(currentMsgid);
-          } else if (isFuzzy) {
-            issues.fuzzyTranslations.push(currentMsgid);
-          } else if (!currentMsgstr || currentMsgstr === "") {
-            issues.untranslated.push(currentMsgid);
-          }
-        }
-
-        // Reset for next entry
-        currentMsgid = "";
-        currentMsgstr = "";
-        isFuzzy = false;
-        isObsolete = false;
-        inMsgid = false;
-        inMsgstr = false;
-
-        if (trimmed.startsWith("msgid ")) {
-          currentMsgid = trimmed.substring(6).replace(/^"(.*)"$/, "$1");
-          inMsgid = true;
-        }
-      }
-    }
-
-    return {
-      package: packageName,
-      locale: locale,
-      file: poFilePath,
-      issues: issues,
-      counts: {
-        untranslated: issues.untranslated.length,
-        fuzzyTranslations: issues.fuzzyTranslations.length,
-        obsoleteTranslations: issues.obsoleteTranslations.length,
-      },
-    };
-  } catch (error) {
-    console.warn(`Failed to validate ${poFilePath}:`, error.message);
-    return null;
-  }
-}
-
-/**
- * Scans an Invenio package directory for translation files and extracts all translations.
- *
- * This function looks for translations in the standard Invenio directory structure:
- * - first tries: packagePath/packageName/translations/
- * - falls back to: packagePath/translations/
- *
- * Within the translations directory, it scans for locale subdirectories containing
- * LC_MESSAGES/messages.po files following the GNU gettext convention.
- *
- * @param {string} packagePath - Path to the package directory
- * @param {string} packageName - Name of the package (used for namespacing)
- * @param {boolean} includeValidation - Whether to include validation analysis
- * @returns {Promise<Object>} Object mapping locale codes to their translations
- *
- * @example
- * // Returns structure like:
- * {
- *   "en": { "Home": "Home", "invenio_app_rdm:Home": "Home" },
- *   "de": { "Home": "Startseite", "invenio_app_rdm:Home": "Startseite" }
- * }
- */
-async function scanPackage(
-  packagePath,
-  packageName,
-  includeValidation = false
-) {
-  const translations = {};
-  const validationReport = [];
-
-  // Find translations directory
-  let translationsDir = path.join(packagePath, packageName, "translations");
-  if (!fs.existsSync(translationsDir)) {
-    translationsDir = path.join(packagePath, "translations");
-  }
-
-  if (!fs.existsSync(translationsDir)) {
-    return { translations, validationReport };
-  }
-
-  // Scan locale directories
-  const locales = fs.readdirSync(translationsDir);
-  for (const locale of locales) {
-    const localePath = path.join(translationsDir, locale);
-    if (fs.statSync(localePath).isDirectory()) {
-      const poFile = path.join(localePath, "LC_MESSAGES/messages.po");
-      if (fs.existsSync(poFile)) {
-        translations[locale] = await parsePoFile(poFile, packageName);
-
-        if (includeValidation) {
-          const validation = createValidationReportOfPoFile(
-            poFile,
-            packageName,
-            locale
-          );
-          if (validation) {
-            validationReport.push(validation);
-          }
-        }
-      }
-    }
-  }
-
-  return { translations, validationReport };
-}
